fix(io): align add-layer decay with compress (float32 + persisted stitching)
Two fixes for the divergence the user reported between:
- /Volumes/.../Dish_1_WT_As_60min.h5 (add-layer flow, broken phasor)
- /Volumes/.../Dish_1_WT_As_FLIM.h5 (compress flow, correct phasor)

Examination of the .h5 files showed the per-pixel decay content was
identical between the two (same brightest-pixel curves, same total
counts, same per-T sums) but tiles landed at different (h, w)
positions — a 90° rotation of one tile-total grid relative to the
other. Compress used its UI defaults (row_by_row / right_down, no
rotation); the user's add-layer dialog used different stitching
choices (snake_by_column / bottom_left / 90° CW). The unfiltered
phasor histogram is rotation-invariant, but the "Filtered" toggle
applies a *spatial* DTCWT smoothing — which is sensitive to neighbor
arrangement, so different tile placements produce different filtered
phasor distributions. That's the visual difference in the screenshots.

Fix 1: persist stitching configuration in /metadata when compress
writes (importer.py). New keys: ``stitch_grid_rows``,
``stitch_grid_cols``, ``stitch_grid_type``, ``stitch_order``. These
let any subsequent re-import / append on the dataset replicate the
exact tile placement compress used — no more remembering Pattern/Start
across two different dialogs.

Fix 2: read those metadata keys in add-layer's TCSPC tab on Scan and
pre-fill the Tile Stitching checkbox + Rows/Cols/Pattern/Start
widgets to mirror compress's choices (gui/add_layer_dialog.py
_tcspc_seed_stitching_from_metadata). Falls back to compress's UI
defaults (row_by_row / right_down, no stitching) when the metadata is
absent. Status-bar message confirms the seed value the user gets.

Fix 3: cast decay arrays to float32 in DatasetStore.append_decay_layers
before write, matching compress's storage format. Phasor math runs in
float64 either way so numerics are identical, but matching dtype keeps
on-disk layout consistent between the two import flows.

Updated test_append_creates_decay_and_provenance to assert float32
dtype (previously expected uint16, the caller's input dtype). 175
tests pass.

+++ tests/test_io/test_store_append.py
@@ -59,7 +59,9 @@ def test_append_creates_decay_and_provenance(tmp_path):
     with h5py.File(store.path, "r") as f:
         assert "decay/ch00" in f
         assert f["decay/ch00"].shape == (32, 32, 16)
-        assert f["decay/ch00"].dtype == np.uint16
+        # Decay is cast to float32 in the store boundary to match compress's
+        # storage format (see store.append_decay_layers docstring).
+        assert f["decay/ch00"].dtype == np.float32
         prov = f["provenance/decay/ch00"].attrs
         assert prov["source_path"] == "/abs/ch00.bin"
         assert prov["cross_format_rule"] == "ZeroPadOffsetRule"
